services/web/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
ecdsa==0.18.0
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Login==0.6.2
frozenlist==1.3.3
hiredis==2.2.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
more-itertools==8.14.0
multidict==6.0.4
numpy==1.24.3
passlib==1.7.4
Pillow==9.5.0
pptree==3.1
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
python-jose==3.3.0
python-ulid==1.1.0
redis==4.5.5
redis-om==0.1.2
requests==2.30.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
types-pyOpenSSL==23.1.0.3
types-redis==4.5.5.2
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.3
yarl==1.9.2
